backend/pyproject.toml
[tool.poetry]
name = "Flask User Authentication"
version = "0.1.0"
description = "Flask application with User Authentication"
authors = ["Pratik Solanki"]

[tool.poetry.dependencies]
python = "^3.12"
flask = "^3.1.1"
flask-sqlalchemy = "^3.1.1"
flask-migrate = "^4.1.0"
flask-jwt-extended = "^4.7.1"
flask-cors = "^6.0.0"
flask-limiter = "^3.12"
psycopg2-binary = "^2.9.10"
redis = "^6.2.0"
celery = "^5.5.3"
gunicorn = "^23.0.0"
marshmallow = "^4.0.0"
marshmallow-sqlalchemy = "^1.4.2"
email-validator = "^2.2.0"
python-dotenv = "^1.1.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.0"
pytest-flask = "^1.3.0"
mypy = "^1.16.0"
black = "^25.1.0"
flake8 = "^7.2.0"
fakeredis = "^2.29.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
